+++ CHANGELOG.md
@@ -13,3 +13,13 @@ All notably significant or outstanding changes to the zlx language will be docum
 - Added support for escape characters in strings
     - Multiline strings have been changed from triple quotes (""") to triple backticks (```)
 - Updated syntax extension and docs to reflect these changes
+
+## [0.2.0]
+- "_" can now be used as an identifier to discard
+    - It cannot be used after the keywords 'const' or 'let'
+- Made the 'typeof' prefix operator more clear, now returns a pair
+    - The 'first' field is a broad type descriptor (e.g. any)
+    - The 'second' field is a narrower representation of the type, aligned with the interpreters inferred or explicit type (e.g. number)
+- Added star star (**) operator that behaves like the same operator in python. It can be used for exponentiation
+    - This is a binary operator
+- Compound assignment now works on more types with more consistent behavior